backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu==2.20.0
numpy
scipy
pyttsx3
pydantic
requests
scikit-learn
